Moved tracing middleware to a sub-package
