+++ internal/server/middleware/tracing/span.go
@@ -1,4 +1,4 @@
-package middleware
+package tracing
 
 // Taken from: middleware/tracing/span.go
